+++ Tests/LanguageServices.Tests.Unit/Declarations/Correct/MachineTests.cs
@@ -61,49 +61,5 @@ class S : MachineState
             Assert.AreEqual(expected.Replace(Environment.NewLine, string.Empty),
                 syntaxTree.ToString().Replace("\n", string.Empty));
         }
-
-        [TestMethod, Timeout(10000)]
-        public void TestNestedMachineDeclaration()
-        {
-            var test = @"
-namespace Foo {
-machine M {
-machine[] MachineArray;
-List<machine> MachineList; 
-List<machine[]> MachineArrayList; 
-start state S { }
-}
-}";
-
-            var configuration = Configuration.Create();
-            configuration.Verbose = 2;
-            
-            var context = CompilationContext.Create(configuration).LoadSolution(test);
-
-            ParsingEngine.Create(context).Run();
-            RewritingEngine.Create(context).Run();
-
-            var syntaxTree = context.GetProjects()[0].PSharpPrograms[0].GetSyntaxTree();
-
-            var expected = @"
-using Microsoft.PSharp;
-namespace Foo
-{
-class M : Machine
-{
-MachineId[] MachineArray;
-List<MachineId> MachineList;
-List<MachineId[]> MachineArrayList;
-[Microsoft.PSharp.Start]
-class S : MachineState
-{
-}
-}
-}";
-
-            Assert.AreEqual(expected.Replace(Environment.NewLine, string.Empty),
-                syntaxTree.ToString().Replace("\n", string.Empty));
-        }
-
     }
 }
